requirements.txt
alembic==1.13.3
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
astroid==3.3.11
backoff==2.2.1
black==25.9.0
boto3==1.35.36
botocore==1.35.99
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dill==0.4.0
distlib==0.4.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.120.1
filelock==3.19.1
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.1.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.9
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.48.0
tenacity==9.1.2
tomli==2.3.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.3
websockets==15.0.1
zipp==3.23.0
